requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alembic==1.0.2            # via flask-migrate
blinker==1.4              # via flask-debugtoolbar
certifi==2018.10.15       # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask
dominate==2.3.4           # via flask-bootstrap4
dropbox==9.1.0
flask-assets==0.12
flask-bootstrap4==4.0.2
flask-caching==1.4.0
flask-debugtoolbar==0.10.1
flask-limiter==1.0.1
flask-login==0.4.1
flask-migrate==2.3.0
flask-sqlalchemy==2.3.2
flask-sslify==0.1.5
flask-uploads==0.2.1
flask-wtf==0.14.2
flask==1.0.2
gunicorn==19.9.0
idna==2.7                 # via requests
itsdangerous==1.1.0       # via flask, flask-debugtoolbar
jinja2==2.10              # via flask
limits==1.3               # via flask-limiter
mako==1.0.7               # via alembic
markupsafe==1.1.0         # via jinja2, mako
pipdeptree==0.13.1
psycopg2-binary==2.7.5
python-dateutil==2.7.5
python-editor==1.0.3      # via alembic
rauth==0.7.3
redis==2.10.6
requests==2.20.0          # via dropbox, rauth
six==1.11.0               # via dropbox, flask-limiter, limits, python-dateutil
sqlalchemy==1.2.13        # via alembic, flask-sqlalchemy
urllib3==1.24.1           # via requests
visitor==0.1.3            # via flask-bootstrap4
webassets==0.12.1         # via flask-assets
werkzeug==0.14.1          # via flask, flask-debugtoolbar
wtforms==2.2.1            # via flask-wtf
xmltodict==0.11.0
